% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{box_base_xpath}
\alias{box_base_xpath}
\title{Find \code{box::use} calls for libraries/packages and local modules}
\usage{
box_base_xpath()
}
\value{
An XPath
}
\description{
Find \code{box::use} calls for libraries/packages and local modules
}
\details{
Base XPath to find \code{box::use} declarations that match the following patterns:
\code{
box::use(
  path/to/module,
)
}

and
\code{
box::use(
  package,
)
}
}
\seealso{
\code{\link[=box_separate_calls_linter]{box_separate_calls_linter()}}
}
\keyword{internal}
