% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{box_default_linters}
\alias{box_default_linters}
\title{Box-compatible default linters}
\format{
An object of class \code{list} of length 34.
}
\usage{
box_default_linters
}
\description{
A replacement for \code{lintr::object_usage_linter()} that works with \code{box} modules.
}
\examples{
linters <- lintr::linters_with_defaults(defaults = box.linters::box_default_linters)

names(linters)

}
\keyword{datasets}
