% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_mod_fun_exists_linter.R
\name{box_mod_fun_exists_linter}
\alias{box_mod_fun_exists_linter}
\title{\code{box} library attached function exists in module linter}
\usage{
box_mod_fun_exists_linter()
}
\description{
\code{box} library attached function exists in module linter
}
