% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{box_module_base_path}
\alias{box_module_base_path}
\title{Find \code{box::use} calls for local modules}
\usage{
box_module_base_path()
}
\value{
An XPath
}
\description{
Find \code{box::use} calls for local modules
}
\details{
Base XPath to find \code{box::use} declarations that match the following pattern:
\code{
box::use(
  path/to/module,
)
}
}
\seealso{
\code{\link[=box_separate_calls_linter]{box_separate_calls_linter()}}
}
\keyword{internal}
