% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_repeated_calls_linter.R
\name{box_repeated_calls_linter}
\alias{box_repeated_calls_linter}
\title{\code{box} library repeated imports of packages}
\usage{
box_repeated_calls_linter()
}
\value{
A custom linter function for use with \code{r-lib/lintr}.
}
\description{
Checks that modules and libraries imports are not repeated.
}
\details{
For use in \code{rhino}, see the
\href{https://appsilon.github.io/rhino/articles/explanation/rhino-style-guide.html}{Explanation: Rhino style guide}
to learn about the details.
}
\examples{
\dontrun{
# will produce lints
lintr::lint(
  text = "box::use(packageA, packageA, packageB)",
  linters = box_repeated_calls_linter()
)

code <- "
box::use(
  dplyr,
  shiny,
)

box::use(
  dplyr,
)
"
lintr::lint(text = code, linters = box_repeated_calls_linter())

code <- "
box::use(
  dplyr[mutate, select],
  dplyr[group_by],
)

box::use(
  path/to/A,
  path/to/A[f1, f2]
)
"
lintr::lint(text = code, linters = box_repeated_calls_linter())

# okay
code <- "
box::use(
  path/to/fileA,
  path/to/fileB,
  path/to/fileC,
)
"
lintr::lint(text = code, linters = box_repeated_calls_linter())

code <- "
box::use(
  path/to/fileA,
  path/to/fileB[f1, f2],
  path/to/fileC[f3, f4],
)
"
lintr::lint(text = code, linters = box_repeated_calls_linter())
}
}
