% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_unused_attached_mod_fun_linter.R
\name{box_unused_attached_mod_obj_linter}
\alias{box_unused_attached_mod_obj_linter}
\title{\code{box} library unused attached module object linter}
\usage{
box_unused_attached_mod_obj_linter()
}
\description{
\code{box} library unused attached module object linter
}
