% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_mod_fun_exists_linter.R
\name{check_attached_mod_funs}
\alias{check_attached_mod_funs}
\title{Check if functions being attached exist and are being exported by the local module}
\usage{
check_attached_mod_funs(xml, xpath)
}
\value{
An XML node list
}
\description{
Check if functions being attached exist and are being exported by the local module
}
\seealso{
\code{\link[=get_module_exports]{get_module_exports()}}
}
\keyword{internal}
