% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_pkg_fun_exists_linter.R
\name{check_attached_pkg_funs}
\alias{check_attached_pkg_funs}
\title{Check if functions being attached exist and are being exported by the library/package}
\usage{
check_attached_pkg_funs(xml, xpath)
}
\value{
An XML node list
}
\description{
Check if functions being attached exist and are being exported by the library/package
}
\seealso{
\code{\link[=get_packages_exports]{get_packages_exports()}}
}
\keyword{internal}
