% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_unique_names_linter.R
\name{cross_duplicated_values}
\alias{cross_duplicated_values}
\title{Identify Cross-Duplicated Values Across Vectors}
\usage{
cross_duplicated_values(vecs)
}
\arguments{
\item{vecs}{A list of atomic vectors. Each element of the list should be a vector
(e.g., character, numeric, etc.) to check for cross-duplicates.}
}
\value{
A list with two elements:
\describe{
\item{flags}{A logical vector indicating whether each vector contains any values that are duplicates of those in previous vectors.}
\item{values}{A list of vectors containing the duplicated values (if any) for each input vector.}
}
}
\description{
This function checks for duplicate values across a list of vectors.
It identifies whether each vector contains values that have already appeared
in any of the previous vectors in the list.
}
\examples{
 \dontrun{
cross_duplicated_values(list(
  c("a", "b", "c"),
  c("d", "b", "e"),
  c("f", "g", "a"),
  c("h", "i", "j")
))
# $flags
# [1] FALSE TRUE TRUE FALSE
#
# $values
# [[1]]
# character(0)
#
# [[2]]
# [1] "b"
#
# [[3]]
# [1] "a"
#
# [[4]]
# character(0)
}

}
\keyword{internal}
