% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_repeated_calls_linter.R
\name{extract_import_parts}
\alias{extract_import_parts}
\title{Extract components before [ bracket}
\usage{
extract_import_parts(target_nodes)
}
\arguments{
\item{target_nodes}{An expression of target_nodes}
}
\description{
Extract components before [ bracket
}
