% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_usage_helper_functions.R
\name{extract_xml_and_text}
\alias{extract_xml_and_text}
\title{Extracts XML nodes and text strings matched by XML search}
\usage{
extract_xml_and_text(xml, xpath)
}
\arguments{
\item{xml}{An XML node list.}

\item{xpath}{An XPath to search xml nodes.}
}
\value{
A list of \code{xml_nodes} and \code{text}.
}
\description{
Extracts XML nodes and text strings matched by XML search
}
\keyword{internal}
