% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_repeated_calls_linter.R
\name{find_all_imports}
\alias{find_all_imports}
\title{Find all imports from an expression}
\usage{
find_all_imports(xml)
}
\arguments{
\item{xml}{A xml coming usually from \code{source_expression$full_xml_parsed_content}
Extract components before [ bracket}
}
\description{
Find all imports from an expression
}
