% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_module_usage_helper_functions.R
\name{get_attached_mod_three_dots}
\alias{get_attached_mod_three_dots}
\title{Get all functions exported from modules by ...}
\usage{
get_attached_mod_three_dots(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
\code{xml} list of \code{xml_nodes}, \code{nested} list of \verb{module: function}, \code{text} list of function
names.
}
\description{
Get all functions exported from modules by ...
}
\keyword{internal}
