% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_module_usage_helper_functions.R
\name{get_attached_modules}
\alias{get_attached_modules}
\title{Get all modules imported whole}
\usage{
get_attached_modules(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
\code{xml} list of \code{xml_nodes}, \code{nested} list of \code{module: functions}, \code{aliases} a named list
of \code{module} = \code{alias}, \code{text} list of all \verb{module$function} names.
}
\description{
Get all modules imported whole
}
\keyword{internal}
