% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_package_usage_helper_functions.R
\name{get_attached_packages}
\alias{get_attached_packages}
\title{Get all packages imported whole}
\usage{
get_attached_packages(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
\code{xml} list of \code{xml_nodes}, \code{nested} list of \code{package: functions}, \code{aliases} a named list
of \code{package} = \code{alias}, \code{text} list of all \verb{package$function} names.
}
\description{
Get all packages imported whole
}
\keyword{internal}
