% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_package_usage_helper_functions.R
\name{get_attached_pkg_functions}
\alias{get_attached_pkg_functions}
\title{Get imported functions from packages}
\usage{
get_attached_pkg_functions(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
\code{xml} list of xml nodes, \code{text} a list of function names.
}
\description{
Get imported functions from packages
}
\keyword{internal}
