% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_package_usage_helper_functions.R
\name{get_attached_pkg_three_dots}
\alias{get_attached_pkg_three_dots}
\title{Get all functions exported from packages by ...}
\usage{
get_attached_pkg_three_dots(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
\code{xml} list of \code{xml_nodes}, \code{nested} list of \verb{package: function}, \code{text} a list of
function names.
}
\description{
Get all functions exported from packages by ...
}
\keyword{internal}
