% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box_module_exports.R
\name{get_box_module_exports}
\alias{get_box_module_exports}
\title{Get a list of functions and data objects exported by a local \code{box} module}
\usage{
get_box_module_exports(declaration, alias = "", caller = globalenv())
}
\arguments{
\item{declaration}{The mod or package call expression. See \code{box:::parse_spec()} source
code for more details.}

\item{alias}{Mod or package-level alias as a character string. See \code{box:::parse_spec()} source
code for more details. Default is fine for \code{box.linters} use.}

\item{caller}{The environment from which \code{box::use} was invoked.
Default is fine for \code{box.linters} use.}
}
\value{
A list of exported functions and data objects.
}
\description{
An implementation for \code{box} modules to provide similar functionality as \code{getNamespaceExports()}
for libraries/packages.
}
\seealso{
\code{\link[=getNamespaceExports]{getNamespaceExports()}}
}
\keyword{internal}
