% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_usage_helper_functions.R
\name{get_declared_objects}
\alias{get_declared_objects}
\title{Get locally declared/defined data objects}
\usage{
get_declared_objects(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
A list of \code{xml_nodes} and \code{text}.
}
\description{
Get locally declared/defined data objects
}
\keyword{internal}
