% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_usage_helper_functions.R
\name{get_function_calls}
\alias{get_function_calls}
\title{Get functions called in current source file}
\usage{
get_function_calls(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
A list of \code{xml_nodes} and \code{text}.
}
\description{
Will not return \verb{package::function()} form. \code{\link[=namespaced_function_calls]{namespaced_function_calls()}} is responsible for
checking \code{package_function()} use.
}
\keyword{internal}
