% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_usage_helper_functions.R
\name{get_function_signature_objs}
\alias{get_function_signature_objs}
\title{Get objects names in function signatures from all functions in current source file}
\usage{
get_function_signature_objs(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
a list of \code{xml_nodes} and \code{text}.
}
\description{
This is a brute-force extraction of \code{SYMBOL_FORMALS} and is not scope-aware.
}
\keyword{internal}
