% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_module_usage_helper_functions.R
\name{get_module_exports}
\alias{get_module_exports}
\title{Get functions exported by modules}
\usage{
get_module_exports(mod_list)
}
\arguments{
\item{mod_list}{A vector of packages}
}
\value{
A list of (\code{package_name} = list(\code{of_functions}))
}
\description{
Get functions exported by modules
}
\keyword{internal}
