% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects_in_strings.R
\name{get_objects_in_strings}
\alias{get_objects_in_strings}
\title{Get objects used in \code{glue} string templates}
\usage{
get_objects_in_strings(xml)
}
\arguments{
\item{xml}{An xml node list.}
}
\value{
A character vector of object and function names found inside \code{glue} string templates.
}
\description{
In \code{glue}, all text between \verb{\{} and \verb{\}} is considered code. Literal braces are defined as
\verb{\{\{} and \verb{\}\}}. Text between double braces are not interpolated.
}
\keyword{internal}
