% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_package_usage_helper_functions.R
\name{get_packages_exports}
\alias{get_packages_exports}
\title{Get functions exported by packages}
\usage{
get_packages_exports(pkg_list)
}
\arguments{
\item{pkg_list}{A vector of packages}
}
\value{
A list of (\code{package_name} = list(\code{of_functions}))
}
\description{
Get functions exported by packages
}
\keyword{internal}
