% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{get_r6_components}
\alias{get_r6_components}
\title{Get declared/defined R6 class components}
\usage{
get_r6_components(xml, mode = c("public", "active", "private"))
}
\arguments{
\item{xml}{XML representation of R source code.}

\item{mode}{Type of internal component (\code{public}, \code{active}, \code{private}).}
}
\value{
List of XML nodes and corresponding text string values of the nodes
}
\description{
Get declared/defined R6 class components
}
\keyword{internal}
