% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{get_r6_inherited_components}
\alias{get_r6_inherited_components}
\title{Get declared/defined R6 class components}
\usage{
get_r6_inherited_components(
  r6_class,
  full_xml,
  source_file = NULL,
  visited = character()
)
}
\arguments{
\item{r6_class}{XML node of the R6Class call being checked.}

\item{full_xml}{Full file XML to search for parent class definitions.}

\item{visited}{Character vector of already-visited class names (cycle guard).}

\item{xml}{XML representation of R source code.}

\item{mode}{Type of internal component (\code{public}, \code{active}, \code{private}).}
}
\value{
List of XML nodes and corresponding text string values of the nodes

Character vector of inherited component names (e.g., "self$log", "private$ctx")
}
\description{
Get declared/defined R6 class components
}
\keyword{=}
\keyword{Get}
\keyword{Looks}
\keyword{ParentClass`}
\keyword{R6}
\keyword{R6Class}
\keyword{Recurses}
\keyword{`inherit}
\keyword{active,}
\keyword{and}
\keyword{call,}
\keyword{class}
\keyword{classes}
\keyword{components}
\keyword{components.}
\keyword{definition}
\keyword{extracts}
\keyword{file,}
\keyword{finds}
\keyword{for}
\keyword{from}
\keyword{grandparents.}
\keyword{in}
\keyword{inherited}
\keyword{internal}
\keyword{its}
\keyword{parent}
\keyword{private}
\keyword{public,}
\keyword{same}
\keyword{the}
