% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_usage_helper_functions.R
\name{get_special_calls}
\alias{get_special_calls}
\title{Get specials called in current source file}
\usage{
get_special_calls(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
A list of \code{xml_nodes} and \code{text}.
}
\description{
Get specials called in current source file
}
\keyword{internal}
