% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{make_r6_components_xpath}
\alias{make_r6_components_xpath}
\title{XPath to get internal components of an R6 class}
\usage{
make_r6_components_xpath(mode = c("public", "active", "private"))
}
\arguments{
\item{mode}{Type of internal component (\code{public}, \code{active}, or \code{private}).}
}
\value{
An XPath query
}
\description{
XPath to get internal components of an R6 class
}
\keyword{internal}
