% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{make_r6_internal_calls_xpath}
\alias{make_r6_internal_calls_xpath}
\title{XPath to get internal function or data object calls inside an R6 class}
\usage{
make_r6_internal_calls_xpath(mode = c("self", "private"))
}
\arguments{
\item{mode}{Type of internal call (\code{self} or \code{private}).}
}
\value{
An XPath query
}
\description{
XPath to get internal function or data object calls inside an R6 class
}
\keyword{internal}
