% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{r6_usage_linter}
\alias{r6_usage_linter}
\title{R6 class usage linter}
\usage{
r6_usage_linter()
}
\value{
A custom linter function for use with \code{r-lib/lintr}.
}
\description{
Checks method and attribute calls within an R6 class. Covers public, private, and active objects.
All internal calls should exist. All private methods and attributes should be used.
}
