% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{resolve_r6_from_module_alias}
\alias{resolve_r6_from_module_alias}
\title{Resolve R6 class from a module alias (inherit = alias$Class)}
\usage{
resolve_r6_from_module_alias(
  class_name,
  module_alias,
  full_xml,
  source_file,
  visited = character()
)
}
\description{
Resolve R6 class from a module alias (inherit = alias$Class)
}
\keyword{internal}
