% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{resolve_r6_from_module_function_import}
\alias{resolve_r6_from_module_function_import}
\title{Resolve R6 class from a function import (box::use(./path[ClassName]))}
\usage{
resolve_r6_from_module_function_import(
  class_name,
  full_xml,
  source_file,
  visited = character()
)
}
\description{
Resolve R6 class from a function import (box::use(./path[ClassName]))
}
\keyword{internal}
