% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{resolve_r6_from_package}
\alias{resolve_r6_from_package}
\title{Load an R6 class from an installed package and extract components}
\usage{
resolve_r6_from_package(class_name, pkg_name)
}
\description{
Load an R6 class from an installed package and extract components
}
\keyword{internal}
