% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{resolve_r6_from_package_import}
\alias{resolve_r6_from_package_import}
\title{Resolve R6 class from a package import (box::use(pkg[ClassName]))}
\usage{
resolve_r6_from_package_import(class_name, full_xml)
}
\description{
Resolve R6 class from a package import (box::use(pkg[ClassName]))
}
\keyword{internal}
