% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_usage_linter.R
\name{resolve_r6_from_same_file}
\alias{resolve_r6_from_same_file}
\title{Find an R6 class in the same file and extract its components}
\usage{
resolve_r6_from_same_file(class_name, full_xml)
}
\description{
Find an R6 class in the same file and extract its components
}
\keyword{internal}
