% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_repeated_calls_linter.R
\name{retrieve_text_before_bracket}
\alias{retrieve_text_before_bracket}
\title{Extract the text before the \code{[} bracket}
\usage{
retrieve_text_before_bracket(text_vector)
}
\arguments{
\item{text_vector}{A character vector}
}
\value{
character vector of text before the first \code{[}
}
\description{
Extract the text before the \code{[} bracket
}
\keyword{internal}
