% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{rhino_default_linters}
\alias{rhino_default_linters}
\title{Rhino default linters}
\format{
An object of class \code{list} of length 38.
}
\usage{
rhino_default_linters
}
\description{
See the \href{https://appsilon.github.io/rhino/articles/explanation/rhino-style-guide.html}{Explanation: Rhino style guide}
to learn about the details.
}
\examples{
linters <- lintr::linters_with_defaults(defaults = box.linters::rhino_default_linters)

names(linters)

}
\keyword{datasets}
