% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_box_use.R
\name{style_box_use_file}
\alias{style_box_use_file}
\title{Style the box::use() calls of a source code}
\usage{
style_box_use_file(filename, run_styler = FALSE)
}
\arguments{
\item{filename}{A file to style.}

\item{run_styler}{Boolean to run styler::style_file() at the end of \code{box::use()} styling.}
}
\description{
Style the box::use() calls of a source code
}
