% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_box_use.R
\name{style_box_use_text}
\alias{style_box_use_text}
\title{Style the box::use() calls of source code text}
\usage{
style_box_use_text(text, indent_spaces = 2)
}
\arguments{
\item{text}{Source code in text format}

\item{indent_spaces}{Number of spaces per indent level}
}
\description{
Style the box::use() calls of source code text
}
