% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_declared_func_linter.R
\name{unused_declared_func_linter}
\alias{unused_declared_func_linter}
\title{Unused declared function linter}
\usage{
unused_declared_func_linter()
}
\description{
Unused declared function linter
}
