% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused_declared_object_linter.R
\name{unused_declared_object_linter}
\alias{unused_declared_object_linter}
\title{Unused declared function and objects linter}
\usage{
unused_declared_object_linter()
}
\description{
Unused declared function and objects linter
}
