% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_repeated_calls_linter.R
\name{find_all_imports}
\alias{find_all_imports}
\title{Get all modules and packages called as a whole, functions, or three-dots}
\usage{
find_all_imports(xml)
}
\arguments{
\item{xml}{An XML node list}
}
\value{
\code{xml} list of \code{xml_nodes}, \code{text} list of module/package names.
}
\description{
Get all modules and packages called as a whole, functions, or three-dots
}
\keyword{internal}
