% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_repeated_calls_linter.R
\name{find_all_imports}
\alias{find_all_imports}
\title{Find all imports from an expression}
\usage{
find_all_imports(box_use_calls)
}
\arguments{
\item{box_use_calls}{A xml coming from \code{xml2::xml_find_all(xml, xpath_base)}
Extract components before [ bracket}
}
\description{
Find all imports from an expression
}
