% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_box_use.R
\name{is_treesitter_installed}
\alias{is_treesitter_installed}
\title{Check if treesitter and dependencies are installed}
\usage{
is_treesitter_installed()
}
\value{
Logical TRUE/FALSE if the \code{treesitter} dependencies exist.
}
\description{
Treesitter required R >= 4.3.0. Treesitter is required by a few \code{{box.linters}} functions.
}
\examples{
\dontrun{

# Bare environment

is_treesitter_installed()
#> [1] FALSE

install.packages(c("treesitter", "treesitter.r"))
is_treesitter_installed()
#> [1] TRUE
}

}
